Replace error message boxes and add sys.stderr to GUI output
