+++ clickqt/core/core.py
@@ -2,7 +2,7 @@ import click
 import inspect
 from clickqt.widgets.multivaluewidget import MultiValueWidget
 from PySide6.QtWidgets import QApplication, QWidget, QVBoxLayout, QPushButton, QTabWidget, QMessageBox, QPlainTextEdit
-from PySide6.QtGui import QTextCursor
+from PySide6.QtGui import QColor
 from clickqt.widgets.checkbox import CheckBox
 from clickqt.widgets.textfield import TextField
 from clickqt.widgets.passwordfield import PasswordField
@@ -13,9 +13,9 @@ from clickqt.widgets.tuplewidget import TupleWidget
 from clickqt.widgets.pathfield import PathField
 from clickqt.widgets.filefield import FileFild
 from clickqt.core.error import ClickQtError
+from clickqt.core.output import Output
 from typing import Dict, Callable, List, Any, Tuple
 import sys
-from io import BytesIO, TextIOWrapper
 
 def qtgui_from_click(cmd):
     # Command-name to command-options and callables
@@ -97,7 +97,7 @@ def qtgui_from_click(cmd):
     def check_error(err: ClickQtError) -> bool:
         if err != ClickQtError.NO_ERROR:
             if err != ClickQtError.ABORTED_ERROR:
-                QMessageBox(QMessageBox.Information, "Error", str(err), QMessageBox.Ok).exec()
+                print(str(err), file=sys.stderr)
             return True
         return False
     
@@ -180,25 +180,12 @@ def qtgui_from_click(cmd):
     run_button.clicked.connect(run)
     layout.addWidget(run_button)
 
-    class Output(TextIOWrapper):
-        """
-            Redirects a stream (here: stdout) to a QPlainTextEdit   
-        """
-        def __init__(self, output: QPlainTextEdit):
-            super().__init__(BytesIO(), sys.stdout.encoding)
-            self.output = output
-
-        def write(self, message: bytes|str):
-            if message:
-                message = message.decode(sys.stdout.encoding) if isinstance(message, bytes) else message  
-                self.output.moveCursor(QTextCursor.End)
-                self.output.insertPlainText(message)
-
     terminal_output = QPlainTextEdit()
     terminal_output.setReadOnly(True)
     terminal_output.setToolTip("Terminal output")
     layout.addWidget(terminal_output)
     sys.stdout = Output(terminal_output)
+    sys.stderr = Output(terminal_output, QColor("red"))
 
     def run_app():
         window.show()
